send-mail-flask/requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.11.0
starkbank-ecdsa==2.2.0
Werkzeug==3.1.3